+++ crates/matrix-sdk-base/src/room/mod.rs
@@ -155,6 +155,13 @@ impl Room {
         self.info.read().room_type().is_some_and(|t| *t == RoomType::Space)
     }
 
+    /// Whether this room is a Call room as defined by [MSC3417].
+    ///
+    /// [MSC3417]: <https://github.com/matrix-org/matrix-spec-proposals/pull/3417>
+    pub fn is_call(&self) -> bool {
+        self.info.read().room_type().is_some_and(|t| *t == RoomType::Call)
+    }
+
     /// Returns the room's type as defined in its creation event
     /// (`m.room.create`).
     pub fn room_type(&self) -> Option<RoomType> {